+++ packages/moonstone/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact moonstone module, newest changes on the top.
 
+## [] -
+
+### Added
+
+- `moonstone/GridListImageItem` support voice control feature
+
 ## [2.1.4] - 2018-09-17
 
 ### Fixed